Applied patch from smizuno #1313
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@9330 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/qgsscalecalculator.cpp
@@ -81,7 +81,7 @@ double QgsScaleCalculator::calculate( const QgsRect &mapExtent, int canvasWidth 
     QgsDebugMsg( "Can't calculate scale from the input values" );
     return 0;
   }
-  double scale = ( delta * conversionFactor ) / ( canvasWidth / mDpi );
+  double scale = ( delta * conversionFactor ) / ( (double)canvasWidth / mDpi );
   return scale;
 }
